fix: change report name
(cherry picked from commit e0fa8904fe28171a9d34781d7c84a97451a5e773)
