+++ india_compliance/gst_india/report/hsn_wise_summary_of_inward_supplies/hsn_wise_summary_of_inward_supplies.json
@@ -2,33 +2,33 @@
  "add_total_row": 1,
  "add_translate_data": 0,
  "columns": [],
- "creation": "2025-04-11 10:32:49.008771",
+ "creation": "2025-04-11 19:47:29.046936",
  "disabled": 0,
  "docstatus": 0,
  "doctype": "Report",
  "filters": [],
  "idx": 0,
  "is_standard": "Yes",
- "letter_head": "",
+ "letter_head": "SMC - New",
  "letterhead": null,
- "modified": "2025-04-11 10:33:30.440910",
+ "modified": "2025-04-11 19:47:29.046936",
  "modified_by": "Administrator",
  "module": "GST India",
- "name": "HSN-wise-summary of purchases",
+ "name": "HSN-wise-summary of inward supplies",
  "owner": "Administrator",
  "prepared_report": 0,
  "ref_doctype": "Purchase Invoice",
- "report_name": "HSN-wise-summary of purchases",
+ "report_name": "HSN-wise-summary of inward supplies",
  "report_type": "Script Report",
  "roles": [
-  {
-   "role": "Accounts Manager"
-  },
   {
    "role": "Accounts User"
   },
   {
    "role": "Auditor"
+  },
+  {
+   "role": "Accounts Manager"
   }
  ],
  "timeout": 0
